requirements.txt
# 项目依赖
python-dotenv==1.0.0
requests==2.31.0